Fix setup for installing npm package for web-listener-runner

+++ web-listener/installer.sh
@@ -15,13 +15,17 @@ if ! id -u web-listener-updater >/dev/null 2>&1; then
   sudo useradd --create-home --user-group web-listener-updater;
 fi;
 if ! id -u web-listener-runner >/dev/null 2>&1; then
-  sudo useradd --system --user-group web-listener-runner;
+  sudo useradd --create-home --system --user-group web-listener-runner;
+  sudo -u web-listener-runner -H sh -c 'mkdir -p ~/.local/bin';
 fi;
 
 # Load dependencies
 
-sudo DEBIAN_FRONTEND=noninteractive apt-get install -y nodejs;
-sudo -u web-listener-runner -H npm install -g web-listener;
+if ! test -f /home/web-listener-runner/.local/bin/web-listener >/dev/null; then
+  sudo DEBIAN_FRONTEND=noninteractive apt-get install -y nodejs;
+  sudo -u web-listener-runner -H npm config set prefix '~/.local/';
+fi;
+sudo -u web-listener-runner -H npm install -g --ignore-scripts web-listener@1.3.2;
 
 # Install boilerplate
